Changing frisbee rpc to FRPC
