--- protoc-gen-frisbee/examples/simple/simple.proto
@@ -1,27 +1,0 @@
-syntax = "proto3";
-
-option go_package = "/simple";
-
-service SimpleService {
-  rpc Echo(Request) returns (Response);
-}
-
-message Request {
-  string Message = 1;
-  bytes B = 2;
-  repeated string Repeat = 3;
-  Response Embedded = 4;
-  message Nested {
-    string N = 1;
-  }
-  Nested NestedM = 5;
-}
-
-message Response{
-  string Message = 1;
-}
-
-message SliceResponse {
-  repeated Response ResponseSlice = 1;
-  map<string, Response> ResponseMap = 2;
-}
