+++ protoc-gen-frpc/examples/pubsub/pubsub.proto
