+++ protoc-gen-frpc/main.go
@@ -14,14 +14,39 @@
 	limitations under the License.
 */
 
-package generator
+package main
 
-const (
-	extension     = ".frisbee.go"
-	pointer       = "*"
-	space         = " "
-	comma         = ","
-	mapSuffix     = "Map"
-	slice         = "[]"
-	packetAnyKind = "packet.AnyKind"
+import (
+	"github.com/loopholelabs/frisbee/protoc-gen-frpc/pkg/generator"
+	"io/ioutil"
+	"os"
 )
+
+func main() {
+	gen := generator.New()
+
+	data, err := ioutil.ReadAll(os.Stdin)
+	if err != nil {
+		panic(err)
+	}
+
+	req, err := gen.UnmarshalRequest(data)
+	if err != nil {
+		panic(err)
+	}
+
+	res, err := gen.Generate(req)
+	if err != nil {
+		panic(err)
+	}
+
+	data, err = gen.MarshalResponse(res)
+	if err != nil {
+		panic(err)
+	}
+
+	_, err = os.Stdout.Write(data)
+	if err != nil {
+		panic(err)
+	}
+}
