+++ protoc-gen-frpc/templates/enums.templ
