+++ protoc-gen-frpc/templates/interfaces.templ
