+++ protoc-gen-frpc/templates/structs.templ
